a lot of refactoring :)
