+++ src/main/java/com/github/lukaszbudnik/dqueue/FunctionalInterfaces.java
@@ -1,0 +1,11 @@
+package com.github.lukaszbudnik.dqueue;
+
+@FunctionalInterface
+interface NoArgFunction<R> {
+    R apply() throws Exception;
+}
+
+@FunctionalInterface
+interface NoArgVoidFunction<R> {
+    void apply() throws Exception;
+}
