+++ src/test/resources/dqueue.properties
@@ -13,3 +13,4 @@ dqueue.cassandraKeyspace=test
 dqueue.cassandraAddress=${env.containsKey('TRAVIS') ? 'localhost' : '192.168.99.100'}
 dqueue.cassandraTablePrefix=dqueue
 dqueue.cassandraCreateTables=true
+dqueue.threadPoolSize=200
