--- src/test/resources/dqueue_guice_module_test.properties
@@ -1,17 +1,0 @@
-#
-# Copyright (C) 2015 Łukasz Budnik <lukasz.budnik@gmail.com>
-#
-# Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with the License. You may obtain a copy of the License at
-#
-# http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the specific language governing permissions and limitations under the License.
-#
-
-# EL test: properties == java.lang.System.getProperties()
-dqueue.cassandraKeyspace=${properties['java.version']}
-# EL test: env == java.lang.System.getenv()
-dqueue.cassandraTablePrefix=${env['HOME']}
-dqueue.cassandraPort=11111
-dqueue.cassandraAddress=cassandra1.docker,cassandra2.docker
-dqueue.cassandraCreateTables=false
